fix bugs

+++ app/schemas.py
@@ -1,19 +1,19 @@
 from pydantic import BaseModel, Field, field_validator
-
+from decimal import Decimal
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount:float
+    amount:Decimal
     description: str|None = Field(None, max_length= 255)
 
     @field_validator('amount')
-    def amount_must_be_positive(cls, v:float)-> float:
+    def amount_must_be_positive(cls, v:Decimal)-> Decimal:
         if v <= 0:
             raise ValueError("Amount must be positive")
         return v
 
     @field_validator('wallet_name')
-    def wallet_name_not_empty(cls, v:str) ->str:
+    def wallet_name_not_empty(cls, v:str) -> str:
         v = v.strip()
         if not v:
             raise ValueError("Wallet name cannot be empty")
@@ -21,7 +21,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name : str = Field(... , max_length = 127)
-    initial_balance : float = 0
+    initial_balance : Decimal = 0
 
     @field_validator('name')
     def name_not_empty(cls, v:str) -> str:
@@ -30,7 +30,7 @@ class CreateWalletRequest(BaseModel):
         return v
 
     @field_validator('initial_balance')
-    def balance_not_negative(cls, v:float) -> float:
+    def balance_not_negative(cls, v:Decimal) -> Decimal:
         if v < 0:
             raise ValueError('Initial balance cannot be negative')
         return v
